Update generate stats imports for relocated sources
